requirements.txt
streamlit>=1.34
pandas>=2.1
scikit-learn>=1.4
joblib>=1.3
numpy>=1.26
python-dotenv>=1.0
google-genai>=0.3.0
matplotlib